limit public queries
