+++ src/typeDefs.ts
@@ -1,6 +1,7 @@
 import path from 'path';
 import fs from 'fs';
 import { gql } from 'apollo-server';
+import { CollectionWithAuthorsAndStories } from './database/queries';
 
 const sharedSchema = fs
   .readFileSync(path.join(__dirname, '..', 'schema-shared.graphql'))
@@ -19,3 +20,16 @@ export const typeDefsAdmin = gql(
     .toString()
     .concat(sharedSchema)
 );
+
+/**
+ * The shape of the response for searchCollections and getCollections queries
+ */
+export type CollectionsResult = {
+  pagination: {
+    totalResults: number;
+    totalPages: number;
+    currentPage: number;
+    perPage: number;
+  };
+  collections: CollectionWithAuthorsAndStories[];
+};
